+++ client/src/commonMain/kotlin/org/apptank/horus/client/sync/manager/SynchronizatorManager.kt
@@ -398,13 +398,20 @@ internal class SynchronizatorManager(
         when (actions) {
             is DataResult.Success -> {
 
-                val newActions =
-                    filterOwnActions(actions.data.map { it.toDomain() }, checkpointDatetime)
+                val newActions = filterOwnActions(actions.data.map { it.toDomain() }, checkpointDatetime)
 
-                val (actionsInsert, updateActions, deleteActions) = organizeActions(newActions)
+                val (moveActions, insertActions, updateActions, deleteActions) = organizeActions(newActions)
 
-                val operations = mapToInsertOperation(actionsInsert) + mapToUpdateOperation(updateActions) + mapToDeleteOperation(deleteActions)
+                if (executeMoveActions(moveActions).not()) {
+                    log("[SynchronizatorManager:synchronizeData] Error executing updelete actions")
+                    syncControlDatabaseHelper.addSyncTypeStatus(
+                        SyncControl.OperationType.CHECKPOINT,
+                        SyncControl.Status.FAILED
+                    )
+                    return false
+                }
 
+                val operations = mapToInsertOperation(insertActions) + mapToUpdateOperation(updateActions) + mapToDeleteOperation(deleteActions)
                 val result = operationDatabaseHelper.executeOperations(operations)
 
                 val syncControlStatus = if (result) {
@@ -593,13 +600,134 @@ internal class SynchronizatorManager(
      * @param syncActions A list of synchronization actions.
      * @return A triple containing lists of insert, update, and delete actions.
      */
-    private fun organizeActions(syncActions: List<SyncControl.Action>): Triple<List<SyncControl.Action>, List<SyncControl.Action>, List<SyncControl.Action>> {
-        val insertActions = syncActions.filter { it.action == SyncControl.ActionType.INSERT }
-            .sortedBy { it.getActionedAtTimestamp() }
-        val updateActions = syncActions.filter { it.action == SyncControl.ActionType.UPDATE }
-            .sortedBy { it.getActionedAtTimestamp() }
-        val deleteActions = syncActions.filter { it.action == SyncControl.ActionType.DELETE }
-            .sortedBy { it.getActionedAtTimestamp() }
-        return Triple(insertActions, updateActions, deleteActions)
+    private fun organizeActions(syncActions: List<SyncControl.Action>): List<List<SyncControl.Action>> {
+        val moveActions = syncActions.filter { it.action == SyncControl.ActionType.MOVE }.sortedBy { it.getActionedAtTimestamp() }
+
+        val insertActions =
+            filterMoveActions(syncActions.filter { it.action == SyncControl.ActionType.INSERT }.sortedBy { it.getActionedAtTimestamp() }, moveActions)
+        val updateActions =
+            filterMoveActions(syncActions.filter { it.action == SyncControl.ActionType.UPDATE }.sortedBy { it.getActionedAtTimestamp() }, moveActions)
+        val deleteActions =
+            filterMoveActions(syncActions.filter { it.action == SyncControl.ActionType.DELETE }.sortedBy { it.getActionedAtTimestamp() }, moveActions)
+
+        return listOf(moveActions, insertActions, updateActions, deleteActions)
+    }
+
+    private fun executeMoveActions(actions: List<SyncControl.Action>): Boolean {
+
+        if (actions.isEmpty()) return true
+
+        if (actions.any { it.action != SyncControl.ActionType.MOVE }) {
+            log("[SynchronizatorManager:executeMoveActions] Invalid action type in move actions")
+            return false
+        }
+
+        // Try to update records
+        if (operationDatabaseHelper.executeOperations(mapToUpdateOperation(actions))) {
+            return true
+        }
+
+        val entitiesSorted = syncControlDatabaseHelper.getEntityNames()
+            .map { it to syncControlDatabaseHelper.getEntityLevel(it) }
+            .sortedByDescending { it.second }
+
+        val groupedByEntity = actions.groupBy { it.entity }
+
+        groupedByEntity.forEach {
+            val entityToDelete = it.key
+            val ids = it.value.map { action -> action.getEntityId() }
+            deleteEntitiesRelated(entityToDelete, ids, entitiesSorted)
+        }
+
+
+        return true
+    }
+
+    private fun deleteEntitiesRelated(
+        entityToDelete: String,
+        ids: List<String>,
+        entitiesSorted: List<Pair<String, Int>>,
+        entitiesProcessed: List<String> = emptyList()
+    ): Boolean {
+
+        if (ids.isEmpty()) return true
+
+        // ---------------------------------------
+        // DELETE RELATED RECORDS
+        // ---------------------------------------
+
+        entitiesSorted.forEach { entitySorted ->
+
+            val entity = entitySorted.first
+            val entityLevel = entitySorted.second
+
+            if (entitiesProcessed.contains(entity)) return@forEach
+
+            syncControlDatabaseHelper.getEntitiesRelated(entity).forEach { entityRelated ->
+
+                val entityRelatedName = entityRelated.entity
+                val entityRelatedLevel = syncControlDatabaseHelper.getEntityLevel(entityRelated.entity)
+
+                if (entityRelatedLevel < entityLevel) {
+
+                    val entityRelatedAttributes = syncControlDatabaseHelper.getEntityAttributes(entityRelatedName)
+                    val entityRelatedIds = operationDatabaseHelper.queryRecords(
+                        SimpleQueryBuilder(entityRelatedName).select(Horus.Attribute.ID).apply {
+                            entityRelatedAttributes.forEach {
+                                whereIn(it, ids, SQL.LogicOperator.OR)
+                            }
+                        }
+                    ).map { it[Horus.Attribute.ID].toString() }
+
+                    // Delete related records
+
+                    if (entityRelatedIds.isNotEmpty()) {
+                        operationDatabaseHelper.deleteRecords(
+                            entity,
+                            entityRelatedIds.map { id ->
+                                entityRelated.attributesLinked.map { attribute ->
+                                    SQL.WhereCondition(SQL.ColumnValue(attribute, id))
+                                }
+                            }.flatten(),
+                            SQL.LogicOperator.OR
+                        )
+                        // Recursive delete
+                        deleteEntitiesRelated(entityRelatedName, entityRelatedIds, entitiesSorted, entitiesProcessed + entity)
+                    }
+                }
+
+            }
+
+        }
+
+        // ---------------------------------------
+        // DELETE PRIMARY RECORDS
+        // ---------------------------------------
+
+        val deleteResult = operationDatabaseHelper.deleteRecords(
+            entityToDelete,
+            ids.map { id -> SQL.WhereCondition(SQL.ColumnValue(Horus.Attribute.ID, id)) },
+            SQL.LogicOperator.OR
+        )
+
+        if (deleteResult.isFailure) {
+            log("[SynchronizatorManager:executeUpdateOrDeleteActions] Error deleting records for entity: $entityToDelete")
+            return false
+        }
+
+        return true
+    }
+
+    private fun filterMoveActions(actions: List<SyncControl.Action>, moveActions: List<SyncControl.Action>): List<SyncControl.Action> {
+        return actions.filter { action ->
+            moveActions.find { moveAction ->
+                action.data.values.forEach {
+                    if (moveAction.data.values.contains(it)) {
+                        return@find true
+                    }
+                }
+                return@find false
+            } == null
+        }
     }
 }